move crate name function to cargo module
