+++ src/cargo.rs
@@ -1,15 +1,49 @@
 //! Functions for retrieving package data from `cargo`.
 
+use std::path::Path;
+use std::process::Command;
+
 use serde_json;
 
 use error::*;
+
+/// Grab the name of the binary or library from it's `Cargo.toml` file.
+///
+/// ## Arguments
+///
+/// - manifest_path: The path to the location of `Cargo.toml` of the crate being documented
+pub fn crate_name_from_manifest_path(manifest_path: &Path) -> Result<String> {
+    let mut command = Command::new("cargo");
+
+    command
+        .arg("metadata")
+        .arg("--manifest-path")
+        .arg(manifest_path.join("Cargo.toml"))
+        .arg("--no-deps")
+        .arg("--format-version")
+        .arg("1");
+
+    let output = command.output()?;
+
+    if !output.status.success() {
+        return Err(
+            ErrorKind::Cargo(
+                output.status,
+                String::from_utf8_lossy(&output.stderr).into_owned(),
+            ).into(),
+        );
+    }
+
+    let metadata = serde_json::from_slice(&output.stdout)?;
+    crate_name_from_metadata(&metadata)
+}
 
 /// Parse the crate name of the binary or library from crate metadata.
 ///
 /// ## Arguments
 ///
 /// - metadata: The JSON metadata of the crate.
-pub fn crate_name_from_metadata(metadata: &serde_json::Value) -> Result<String> {
+fn crate_name_from_metadata(metadata: &serde_json::Value) -> Result<String> {
     let targets = match metadata["packages"][0]["targets"].as_array() {
         Some(targets) => targets,
         None => return Err(ErrorKind::Json("targets is not an array").into()),
